GH-354: [style] update UI to match agreed mockup
